+++ doc/changelog.md
@@ -6,7 +6,7 @@ This files contains a version history including all changes relevant for semanti
 
 ## Upcoming Changes
 
-*(none)*
+* GenericOAuthOperationsProvider now also support body-credentials when retriving tokens by code